requirements.txt
streamlit==1.42.0
Pillow==10.4.0
torch==2.6.0
torchvision==0.21.0